Move redoLogBinary to SessionPageStore
