+++ h2/src/main/org/h2/command/dml/Set.java
@@ -23,6 +23,7 @@ import org.h2.expression.ValueExpression;
 import org.h2.message.DbException;
 import org.h2.message.Trace;
 import org.h2.mode.DefaultNullOrdering;
+import org.h2.pagestore.db.SessionPageStore;
 import org.h2.result.ResultInterface;
 import org.h2.schema.Schema;
 import org.h2.security.auth.AuthenticatorFactory;
@@ -387,7 +388,11 @@ public class Set extends Prepared {
         }
         case SetTypes.REDO_LOG_BINARY: {
             int value = getIntValue();
-            session.setRedoLogBinary(value == 1);
+            if (session instanceof SessionPageStore) {
+                ((SessionPageStore) session).setRedoLogBinary(value == 1);
+            } else {
+                DbException.getUnsupportedException("MV_STORE + SET REDO_LOG_BINARY");
+            }
             break;
         }
         case SetTypes.REFERENTIAL_INTEGRITY: {
